Move UpCloudConfigEntry to coordinator.py
Avoids cyclic imports and their workarounds.

Thanks-to: @epenet
