+++ homeassistant/components/upcloud/coordinator.py
@@ -4,18 +4,18 @@ from __future__ import annotations
 
 from datetime import timedelta
 import logging
-from typing import TYPE_CHECKING
 
 import upcloud_api
 
+from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_SCAN_INTERVAL
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
-if TYPE_CHECKING:
-    from . import UpCloudConfigEntry
-
 _LOGGER = logging.getLogger(__name__)
+
+
+type UpCloudConfigEntry = ConfigEntry[UpCloudDataUpdateCoordinator]
 
 
 class UpCloudDataUpdateCoordinator(
